+++ boilerplates/pc/react/src/components/Index/js/Index.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import '../less/index.less';
+
+class Index extends React.Component {
+  render() {
+    return (
+      <div className="index">
+        React boilerplate
+      </div>
+    );
+  }
+}
+
+export default Index;