+++ boilerplates/pc/react/src/routes.js
@@ -1,0 +1,7 @@
+import { withRouter } from 'react-router';
+import Index from './components/Index/js/Index';
+
+export default {
+  path: '/',
+  component: withRouter(Index)
+};