+++ boilerplates/pc/react/themes/theme.js
@@ -1,0 +1,7 @@
+module.exports = {
+  'primary-color': '#1DA57A',
+  'link-color': '#1DA57A',
+  // 'border-radius-base': 0,
+  'layout-body-background': '#ececec',
+  'layout-header-background': ' #404040',
+};